Add Slink support (and various other minor tweaks)

+++ examples/raspbian.sh
@@ -39,14 +39,21 @@ export TZ='UTC' LC_ALL='C'
 
 dpkgArch='armhf'
 
-#mirror='http://archive.raspbian.org/raspbian'
-mirror='http://mirrordirector.raspbian.org/raspbian'
-# (https://www.raspbian.org/RaspbianMirrors#The_mirror_redirection_system)
-
 exportDir="$tmpDir/output"
 archDir="$exportDir/raspbian/$dpkgArch"
 tmpOutputDir="$archDir/$suite"
 
+#mirror='http://archive.raspbian.org/raspbian'
+mirror='http://mirrordirector.raspbian.org/raspbian'
+# (https://www.raspbian.org/RaspbianMirrors#The_mirror_redirection_system)
+
+initArgs=(
+	--arch "$dpkgArch"
+	--non-debian
+)
+
+export GNUPGHOME="$tmpDir/gnupg"
+mkdir -p "$GNUPGHOME"
 keyring='/usr/share/keyrings/raspbian-archive-keyring.gpg'
 if [ ! -s "$keyring" ]; then
 	# since we're using mirrors, we ought to be more explicit about download verification
@@ -64,30 +71,29 @@ if [ ! -s "$keyring" ]; then
 	gpg --batch --no-default-keyring --keyring "$keyring" --import "$keyring.asc"
 	rm -f "$keyring.asc"
 fi
+initArgs+=( --keyring "$keyring" )
 
 mkdir -p "$tmpOutputDir"
-if wget -O "$tmpOutputDir/InRelease" "$mirror/dists/$suite/InRelease"; then
+
+if [ -f "$keyring" ] && wget -O "$tmpOutputDir/InRelease" "$mirror/dists/$suite/InRelease"; then
 	gpgv \
 		--keyring "$keyring" \
 		--output "$tmpOutputDir/Release" \
 		"$tmpOutputDir/InRelease"
-else
+	[ -s "$tmpOutputDir/Release" ]
+elif [ -f "$keyring" ] && wget -O "$tmpOutputDir/Release.gpg" "$mirror/dists/$suite/Release.gpg" && wget -O "$tmpOutputDir/Release" "$mirror/dists/$suite/Release"; then
 	rm -f "$tmpOutputDir/InRelease" # remove wget leftovers
-	wget -O "$tmpOutputDir/Release.gpg" "$mirror/dists/$suite/Release.gpg"
-	wget -O "$tmpOutputDir/Release" "$mirror/dists/$suite/Release"
 	gpgv \
 		--keyring "$keyring" \
 		"$tmpOutputDir/Release.gpg" \
 		"$tmpOutputDir/Release"
+	[ -s "$tmpOutputDir/Release" ]
+else
+	echo >&2 "error: failed to fetch either InRelease or Release.gpg+Release for '$suite' (from '$mirror')"
+	exit 1
 fi
 
-initArgs=(
-	--arch "$dpkgArch"
-
-	--non-debian
-
-	--keyring "$keyring"
-
+initArgs+=(
 	# disable merged-usr (for now?) due to the following compelling arguments:
 	#  - https://bugs.debian.org/src:usrmerge ("dpkg-query" breaks, etc)
 	#  - https://bugs.debian.org/914208 ("buildd" variant disables merged-usr still)
@@ -112,19 +118,30 @@ touch_epoch() {
 	done
 }
 
-debuerreotype-apt-get "$rootfsDir" dist-upgrade -yqq
+aptVersion="$("$debuerreotypeScriptsDir/.apt-version.sh" "$rootfsDir")"
+if dpkg --compare-versions "$aptVersion" '>=' '1.1~'; then
+	debuerreotype-apt-get "$rootfsDir" full-upgrade -yqq
+else
+	debuerreotype-apt-get "$rootfsDir" dist-upgrade -yqq
+fi
 
 # copy the rootfs to create other variants
 mkdir "$rootfsDir"-slim
 tar -cC "$rootfsDir" . | tar -xC "$rootfsDir"-slim
 
-# prefer iproute2 if it exists
-iproute=iproute2
-if ! debuerreotype-apt-get "$rootfsDir" install -qq -s iproute2 &> /dev/null; then
-	# poor wheezy
-	iproute=iproute
+# for historical reasons (related to their usefulness in debugging non-working container networking in container early days before "--network container:xxx"), Debian 10 and older non-slim images included both "ping" and "ip" above "minbase", but in 11+ (Bullseye), that will no longer be the case and we will instead be a faithful minbase again :D
+epoch2021="$(date --date '2021-01-01 00:00:00' +%s)"
+if [ "$epoch" -lt "$epoch2021" ] || { isDebianBusterOrOlder="$([ -f "$rootfsDir/etc/os-release" ] && source "$rootfsDir/etc/os-release" && [ -n "${VERSION_ID:-}" ] && [ "${VERSION_ID%%.*}" -le 10 ] && echo 1)" && [ -n "$isDebianBusterOrOlder" ]; }; then
+	# prefer iproute2 if it exists
+	iproute=iproute2
+	if ! debuerreotype-apt-get "$rootfsDir" install -qq -s iproute2 &> /dev/null; then
+		# poor wheezy
+		iproute=iproute
+	fi
+	ping=iputils-ping
+	noInstallRecommends='--no-install-recommends'
+	debuerreotype-apt-get "$rootfsDir" install -y $noInstallRecommends $ping $iproute
 fi
-debuerreotype-apt-get "$rootfsDir" install -y --no-install-recommends iputils-ping $iproute
 
 debuerreotype-slimify "$rootfsDir"-slim
 
@@ -152,8 +169,10 @@ create_artifacts() {
 
 	for f in debian_version os-release apt/sources.list; do
 		targetFile="$targetBase.$(basename "$f" | sed -r "s/[^a-zA-Z0-9_-]+/-/g")"
-		cp "$rootfs/etc/$f" "$targetFile"
-		touch_epoch "$targetFile"
+		if [ -e "$rootfs/etc/$f" ]; then
+			cp "$rootfs/etc/$f" "$targetFile"
+			touch_epoch "$targetFile"
+		fi
 	done
 }
